pyproject.toml
[tool.poetry]
name = "akshay-cv-backend"
version = "0.1.0"
description = "Flask backend API for Akshay Sayar's CV website with Gemini LLM integration"
authors = ["Akshay Sayar <akshay.sayar@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-cors = "^4.0.0"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
